refactor: refined scripts
