+++ scripts/gallery.js
@@ -2,8 +2,6 @@ import { refs } from './refs';
 
 const { masonryContainer, expandButton } = refs();
 
-const galleryImages = masonryContainer.querySelectorAll('img');
-
 let isExpanded = false;
 
 expandButton.addEventListener('click', (evt) => {
@@ -16,18 +14,9 @@ expandButton.addEventListener('click', (evt) => {
 
     expandButton.innerHTML = 'Schowaj &uarr;';
     expandButton.style.bottom = '0';
-
-    galleryImages.forEach((image) => image.addEventListener('click', handleImageClick));
   } else {
     masonryContainer.classList.remove('expanded');
     expandButton.innerHTML = 'Rozwiń &darr;';
     expandButton.style.bottom = '6%';
-
-    galleryImages.forEach((image) => image.removeEventListener('click', handleImageClick));
   }
 });
-
-const handleImageClick = (evt) => {
-  const clickedImage = evt.target;
-  console.log(clickedImage.alt);
-};
