+++ scripts/helpers/onMenuActions.js
@@ -1,0 +1,9 @@
+import { refs } from '../refs';
+
+const { mobileMenu } = refs();
+
+const toggleMenu = () => {
+  mobileMenu.classList.toggle('hidden');
+};
+
+export { toggleMenu };
