+++ scripts/slider.js
@@ -1,25 +1,7 @@
-const slides = document.querySelectorAll('[slide]');
+import { showNextSlide, showPrevSlide } from './helpers/onSliderActions';
+
 const prevButton = document.querySelector('[data-prev-slide]');
 const nextButton = document.querySelector('[data-next-slide]');
-let currentIndex = 0;
-
-function showSlide(index) {
-  slides.forEach((slide) => {
-    slide.classList.add('hidden');
-  });
-  slides[index].classList.remove('hidden');
-  currentIndex = index;
-}
-
-function showNextSlide() {
-  currentIndex = (currentIndex + 1) % slides.length;
-  showSlide(currentIndex);
-}
-
-function showPrevSlide() {
-  currentIndex = (currentIndex - 1 + slides.length) % slides.length;
-  showSlide(currentIndex);
-}
 
 prevButton.addEventListener('click', showPrevSlide);
 nextButton.addEventListener('click', showNextSlide);
