+++ scripts/toggleMenu.js
@@ -1,8 +1,19 @@
 import { refs } from './refs';
-const { mobileMenu } = refs();
+import { hideDropdown } from './helpers/onDropdownMenuClick';
+import { hideSearchInput } from './helpers/onSearchInputActions';
+import { toggleMenu } from './helpers/onMenuActions';
 
-const toggleMenu = () => {
-  mobileMenu.classList.toggle('hidden');
-};
+const { menuOpenBtn, menuCloseBtn, mobileMenuButtons } = refs();
+
+menuOpenBtn.addEventListener('click', toggleMenu);
+menuCloseBtn.addEventListener('click', toggleMenu);
+
+mobileMenuButtons.forEach((button) =>
+  button.addEventListener('click', () => {
+    toggleMenu();
+    hideDropdown();
+    hideSearchInput();
+  })
+);
 
 export { toggleMenu };
